extracted the RelaySocket GenServer from its interface
